+++ src/lib/Makefile.inc
@@ -22,6 +22,7 @@ CSOURCES = ares__addrinfo2hostent.c	\
   ares_data.c				\
   ares_destroy.c			\
   ares_dns_mapping.c			\
+  ares_dns_name.c			\
   ares_dns_parse.c			\
   ares_dns_record.c			\
   ares_dns_write.c			\
