+++ src/lib/ares__buf.c
@@ -45,27 +45,7 @@ struct ares__buf {
                                        *   SIZE_MAX if not set. */
 };
 
-/* Reserved characters for names that need to be escaped */
-static ares_bool_t is_reservedch(int ch)
-{
-  switch (ch) {
-    case '"':
-    case '.':
-    case ';':
-    case '\\':
-    case '(':
-    case ')':
-    case '@':
-    case '$':
-      return ARES_TRUE;
-    default:
-      break;
-  }
-
-  return ARES_FALSE;
-}
-
-static ares_bool_t ares__isprint(int ch)
+ares_bool_t ares__isprint(int ch)
 {
   if (ch >= 0x20 && ch <= 0x7E) {
     return ARES_TRUE;
@@ -635,72 +615,6 @@ ares_status_t ares__buf_fetch_bytes_into_buf(ares__buf_t *buf,
   return ares__buf_consume(buf, len);
 }
 
-static ares_status_t ares__buf_fetch_dnsname_into_buf(ares__buf_t *buf,
-                                                      ares__buf_t *dest,
-                                                      size_t       len,
-                                                      ares_bool_t  is_hostname)
-{
-  size_t               remaining_len;
-  const unsigned char *ptr = ares__buf_fetch(buf, &remaining_len);
-  ares_status_t        status;
-  size_t               i;
-
-  if (buf == NULL || len == 0 || remaining_len < len) {
-    return ARES_EBADRESP;
-  }
-
-  for (i = 0; i < len; i++) {
-    unsigned char c = ptr[i];
-
-    /* Hostnames have a very specific allowed character set.  Anything outside
-     * of that (non-printable and reserved included) are disallowed */
-    if (is_hostname && !ares__is_hostnamech(c)) {
-      status = ARES_EBADRESP;
-      goto fail;
-    }
-
-    /* NOTE: dest may be NULL if the user is trying to skip the name. validation
-     *       still occurs above. */
-    if (dest == NULL) {
-      continue;
-    }
-
-    /* Non-printable characters need to be output as \DDD */
-    if (!ares__isprint(c)) {
-      unsigned char escape[4];
-
-      escape[0] = '\\';
-      escape[1] = '0' + (c / 100);
-      escape[2] = '0' + ((c % 100) / 10);
-      escape[3] = '0' + (c % 10);
-
-      status = ares__buf_append(dest, escape, sizeof(escape));
-      if (status != ARES_SUCCESS) {
-        goto fail;
-      }
-
-      continue;
-    }
-
-    /* Reserved characters need to be escaped, otherwise normal */
-    if (is_reservedch(c)) {
-      status = ares__buf_append_byte(dest, '\\');
-      if (status != ARES_SUCCESS) {
-        goto fail;
-      }
-    }
-
-    status = ares__buf_append_byte(dest, c);
-    if (status != ARES_SUCCESS) {
-      return status;
-    }
-  }
-
-  return ares__buf_consume(buf, len);
-
-fail:
-  return status;
-}
 
 size_t ares__buf_consume_whitespace(ares__buf_t *buf,
                                     ares_bool_t  include_linefeed)
@@ -849,161 +763,6 @@ ares_status_t ares__buf_set_position(ares__buf_t *buf, size_t idx)
 
   buf->offset = idx;
   return ARES_SUCCESS;
-}
-
-#define ARES_DNS_HEADER_SIZE 12
-
-ares_status_t ares__buf_parse_dns_name(ares__buf_t *buf, char **name,
-                                       ares_bool_t is_hostname)
-{
-  size_t        save_offset = 0;
-  unsigned char c;
-  ares_status_t status;
-  ares__buf_t  *namebuf     = NULL;
-  size_t        label_start = ares__buf_get_position(buf);
-
-  if (buf == NULL) {
-    return ARES_EFORMERR;
-  }
-
-  if (name != NULL) {
-    namebuf = ares__buf_create();
-    if (namebuf == NULL) {
-      status = ARES_ENOMEM;
-      goto fail;
-    }
-  }
-
-  /* XXX: LibraryTest.ExpandName and LibraryTest.ExpandNameFailure are not
-   *      complete DNS messages but rather non-valid minimized snippets to try
-   *      to test the old parser, so we have to turn off this sanity check
-   *      it appears until these test cases are rewritten
-   */
-#if 0
-  if (ares__buf_get_position(buf) < ARES_DNS_HEADER_SIZE) {
-    status = ARES_EFORMERR;
-    goto fail;
-  }
-#endif
-  /* The compression scheme allows a domain name in a message to be
-   * represented as either:
-   *
-   * - a sequence of labels ending in a zero octet
-   * - a pointer
-   * - a sequence of labels ending with a pointer
-   */
-  while (1) {
-    /* Keep track of the minimum label starting position to prevent forward
-     * jumping */
-    if (label_start > ares__buf_get_position(buf)) {
-      label_start = ares__buf_get_position(buf);
-    }
-
-    status = ares__buf_fetch_bytes(buf, &c, 1);
-    if (status != ARES_SUCCESS) {
-      goto fail;
-    }
-
-    /* Pointer/Redirect */
-    if ((c & 0xc0) == 0xc0) {
-      /* The pointer takes the form of a two octet sequence:
-       *
-       *   +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
-       *   | 1  1|                OFFSET                   |
-       *   +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
-       *
-       * The first two bits are ones.  This allows a pointer to be distinguished
-       * from a label, since the label must begin with two zero bits because
-       * labels are restricted to 63 octets or less.  (The 10 and 01
-       * combinations are reserved for future use.)  The OFFSET field specifies
-       * an offset from the start of the message (i.e., the first octet of the
-       * ID field in the domain header).  A zero offset specifies the first byte
-       * of the ID field, etc.
-       */
-      size_t offset = (size_t)((c & 0x3F) << 8);
-
-      /* Fetch second byte of the redirect length */
-      status = ares__buf_fetch_bytes(buf, &c, 1);
-      if (status != ARES_SUCCESS) {
-        goto fail;
-      }
-
-      offset |= ((size_t)c);
-
-      /* According to RFC 1035 4.1.4:
-       *    In this scheme, an entire domain name or a list of labels at
-       *    the end of a domain name is replaced with a pointer to a prior
-       *    occurance of the same name.
-       * Note the word "prior", meaning it must go backwards.  This was
-       * confirmed via the ISC BIND code that it also prevents forward
-       * pointers.
-       */
-      if (offset >= label_start) {
-        status = ARES_EBADNAME;
-        goto fail;
-      }
-
-      /* First time we make a jump, save the current position */
-      if (save_offset == 0) {
-        save_offset = ares__buf_get_position(buf);
-      }
-
-      status = ares__buf_set_position(buf, offset);
-      if (status != ARES_SUCCESS) {
-        status = ARES_EBADNAME;
-        goto fail;
-      }
-
-      continue;
-    } else if ((c & 0xc0) != 0) {
-      /* 10 and 01 are reserved */
-      status = ARES_EBADNAME;
-      goto fail;
-    } else if (c == 0) {
-      /* termination via zero octet*/
-      break;
-    }
-
-    /* New label */
-
-    /* Labels are separated by periods */
-    if (ares__buf_len(namebuf) != 0 && name != NULL) {
-      status = ares__buf_append_byte(namebuf, '.');
-      if (status != ARES_SUCCESS) {
-        goto fail;
-      }
-    }
-
-    status = ares__buf_fetch_dnsname_into_buf(buf, namebuf, c, is_hostname);
-    if (status != ARES_SUCCESS) {
-      goto fail;
-    }
-  }
-
-  /* Restore offset read after first redirect/pointer as this is where the DNS
-   * message continues */
-  if (save_offset) {
-    ares__buf_set_position(buf, save_offset);
-  }
-
-  if (name != NULL) {
-    *name = ares__buf_finish_str(namebuf, NULL);
-    if (*name == NULL) {
-      status = ARES_ENOMEM;
-      goto fail;
-    }
-  }
-
-  return ARES_SUCCESS;
-
-fail:
-  /* We want badname response if we couldn't parse */
-  if (status == ARES_EBADRESP) {
-    status = ARES_EBADNAME;
-  }
-
-  ares__buf_destroy(namebuf);
-  return status;
 }
 
 ares_status_t ares__buf_parse_dns_binstr(ares__buf_t *buf, size_t remaining_len,
