+++ src/lib/ares_dns_write.c
@@ -32,356 +32,6 @@
 #  include <stdint.h>
 #endif
 
-typedef struct {
-  char  *name;
-  size_t name_len;
-  size_t idx;
-} ares_nameoffset_t;
-
-static void ares__nameoffset_free(void *arg)
-{
-  ares_nameoffset_t *off = arg;
-  if (off == NULL) {
-    return;
-  }
-  ares_free(off->name);
-  ares_free(off);
-}
-
-static ares_status_t ares__nameoffset_create(ares__llist_t **list,
-                                             const char *name, size_t idx)
-{
-  ares_status_t      status;
-  ares_nameoffset_t *off = NULL;
-
-  if (list == NULL || name == NULL || ares_strlen(name) == 0 ||
-      ares_strlen(name) > 255) {
-    return ARES_EFORMERR;
-  }
-
-  if (*list == NULL) {
-    *list = ares__llist_create(ares__nameoffset_free);
-  }
-  if (*list == NULL) {
-    status = ARES_ENOMEM;
-    goto fail;
-  }
-
-  off = ares_malloc_zero(sizeof(*off));
-  if (off == NULL) {
-    return ARES_ENOMEM;
-  }
-
-  off->name     = ares_strdup(name);
-  off->name_len = ares_strlen(off->name);
-  off->idx      = idx;
-
-  if (ares__llist_insert_last(*list, off) == NULL) {
-    status = ARES_ENOMEM;
-    goto fail;
-  }
-
-  status = ARES_SUCCESS;
-
-fail:
-  ares__nameoffset_free(off);
-  return status;
-}
-
-static const ares_nameoffset_t *ares__nameoffset_find(ares__llist_t *list,
-                                                      const char    *name)
-{
-  size_t                   name_len = ares_strlen(name);
-  ares__llist_node_t      *node;
-  const ares_nameoffset_t *longest_match = NULL;
-
-  if (list == NULL || name == NULL || name_len == 0) {
-    return NULL;
-  }
-
-  for (node = ares__llist_node_first(list); node != NULL;
-       node = ares__llist_node_next(node)) {
-    const ares_nameoffset_t *val = ares__llist_node_val(node);
-    size_t                   prefix_len;
-
-    /* Can't be a match if the stored name is longer */
-    if (val->name_len > name_len) {
-      continue;
-    }
-
-    /* Can't be the longest match if our existing longest match is longer */
-    if (longest_match != NULL && longest_match->name_len > val->name_len) {
-      continue;
-    }
-
-    prefix_len = name_len - val->name_len;
-
-    if (strcasecmp(val->name, name + prefix_len) != 0) {
-      continue;
-    }
-
-    /* We need to make sure if `val->name` is "example.com" that name is
-     * is separated by a label, e.g. "myexample.com" is not ok, however
-     * "my.example.com" is, so we look for the preceding "." */
-    if (prefix_len != 0 && name[prefix_len - 1] != '.') {
-      continue;
-    }
-
-    longest_match = val;
-  }
-
-  return longest_match;
-}
-
-typedef struct {
-  unsigned char label[63];
-  size_t        len;
-} ares_dns_label_t;
-
-static ares_status_t ares_parse_dns_name_escape(const char *ptr, size_t ptr_len,
-                                                ares_bool_t validate_hostname,
-                                                unsigned char *out,
-                                                size_t *consumed_chars)
-{
-  /* Must have at least 1 more character */
-  if (ptr_len < 2) {
-    return ARES_EBADNAME;
-  }
-
-  /* If next character is a digit, must have 3 */
-  if (isdigit(ptr[1])) {
-    int  i;
-    char num[4];
-
-    if (ptr_len < 4)
-      return ARES_EBADNAME;
-
-    /* Must all be digits */
-    if (!isdigit(ptr[2]) || !isdigit(ptr[3])) {
-      return ARES_EBADNAME;
-    }
-
-    num[0] = ptr[1];
-    num[1] = ptr[2];
-    num[2] = ptr[3];
-    num[3] = 0;
-    i = atoi(num);
-
-    /* Out of range */
-    if (i > 255)
-      return ARES_EBADNAME;
-
-    if (validate_hostname && !ares__is_hostnamech((unsigned char) i))
-      return ARES_EBADNAME;
-
-    *out            = (unsigned char)i;
-    *consumed_chars = 3;
-    return ARES_SUCCESS;
-  }
-
-  /* We can just output the character */
-  if (validate_hostname && !ares__is_hostnamech(ptr[1]))
-    return ARES_EBADNAME;
-
-  *out            = (unsigned char)ptr[1];
-  *consumed_chars = 1;
-  return ARES_SUCCESS;
-}
-
-static ares_status_t ares_parse_dns_name(ares_dns_label_t **labels_out,
-                                         size_t *num_labels_out,
-                                         ares_bool_t validate_hostname,
-                                         const char *name)
-{
-  ares_status_t     status;
-  ares_dns_label_t *labels     = NULL;
-  ares_dns_label_t *label;
-  size_t            num_labels = 0;
-  size_t            len;
-  size_t            i;
-  size_t            total_len  = 0;
-
-  if (name == NULL || labels_out == NULL || num_labels_out == NULL)
-    return ARES_EFORMERR;
-
-  len = ares_strlen(name);
-
-  /* Start with 1 label */
-  num_labels = 1;
-  labels     = ares_malloc_zero(sizeof(*labels) * num_labels);
-  if (labels == NULL) {
-    status = ARES_ENOMEM;
-    goto done;
-  }
-
-  label = &labels[num_labels - 1];
-  for (i=0; i<len; i++) {
-    size_t remaining_len = len - i;
-
-    /* New label */
-    if (name[i] == '.') {
-      void *temp;
-
-      temp = ares_realloc_zero(labels, sizeof(*labels) * num_labels,
-                               sizeof(*labels) * (num_labels+1));
-      if (temp == NULL) {
-        status = ARES_ENOMEM;
-        goto done;
-      }
-      labels = temp;
-      label  = &labels[num_labels++];
-      continue;
-    }
-
-    /* Can't append any more bytes */
-    if (label->len == sizeof(label->label)) {
-      status = ARES_EBADNAME;
-      goto done;
-    }
-
-    /* Escape */
-    if (name[i] == '\\') {
-      size_t consumed_chars = 0;
-      status = ares_parse_dns_name_escape(name + i, remaining_len,
-        validate_hostname, &label->label[label->len++], &consumed_chars);
-      if (status != ARES_SUCCESS) {
-        goto done;
-      }
-      i += consumed_chars;
-      continue;
-    }
-
-    /* Output direct character */
-    if (validate_hostname && !ares__is_hostnamech(name[i])) {
-      status = ARES_EBADNAME;
-      goto done;
-    }
-    label->label[label->len++] = (unsigned char)name[i];
-  }
-
-  /* Remove trailing blank label */
-  if (labels[num_labels-1].len == 0)
-    num_labels--;
-
-  /* If someone passed in "." there could have been 2 blank labels, check for
-   * that */
-  if (num_labels == 1 && labels[0].len == 0) {
-    num_labels--;
-  }
-
-  /* Scan to make sure there are no blank labels */
-  for (i=0; i<num_labels; i++) {
-    if (labels[i].len == 0) {
-      status = ARES_EBADNAME;
-      goto done;
-    }
-    total_len += labels[i].len;
-  }
-
-  /* Can't exceed maximum (unescaped) length */
-  if (num_labels && total_len + num_labels - 1 > 255) {
-    status = ARES_EBADNAME;
-    goto done;
-  }
-
-  *labels_out     = labels;
-  *num_labels_out = num_labels;
-  status = ARES_SUCCESS;
-
-done:
-  if (status != ARES_SUCCESS)
-    ares_free(labels);
-  return status;
-}
-
-
-static ares_status_t ares_dns_write_name(ares__buf_t *buf, ares__llist_t **list,
-                                         ares_bool_t validate_hostname,
-                                         const char *name)
-{
-  const ares_nameoffset_t *off      = NULL;
-  size_t                   name_len;
-  size_t                   pos      = ares__buf_get_position(buf);
-  ares_dns_label_t        *labels   = NULL;
-  char                     name_copy[512];
-  size_t                   num_labels = 0;
-  ares_status_t            status;
-
-  if (buf == NULL || name == NULL) {
-    return ARES_EFORMERR;
-  }
-
-  /* NOTE: due to possible escaping, name_copy buffer is > 256 to allow for
-   *       this */
-  name_len = ares_strcpy(name_copy, name, sizeof(name_copy));
-
-  /* Find longest match */
-  if (list != NULL) {
-    off = ares__nameoffset_find(*list, name_copy);
-    if (off != NULL && off->name_len != name_len) {
-      /* truncate */
-      name_len                -= (off->name_len + 1);
-      name_copy[name_len - 1]  = 0;
-    }
-  }
-
-  /* Output labels */
-  if (off == NULL || off->name_len != name_len) {
-    size_t i;
-
-    status = ares_parse_dns_name(&labels, &num_labels, validate_hostname,
-                                 name_copy);
-    if (status != ARES_SUCCESS) {
-      goto done;
-    }
-
-    for (i = 0; i < num_labels; i++) {
-      status = ares__buf_append_byte(buf,
-        (unsigned char)(labels[i].len & 0xFF));
-      if (status != ARES_SUCCESS) {
-        goto done;
-      }
-
-      status = ares__buf_append(buf, labels[i].label, labels[i].len);
-      if (status != ARES_SUCCESS) {
-        goto done;
-      }
-    }
-
-    /* If we are NOT jumping to another label, output terminator */
-    if (off == NULL) {
-      status = ares__buf_append_byte(buf, 0);
-      if (status != ARES_SUCCESS) {
-        goto done;
-      }
-    }
-  }
-
-  /* Output name compression offset jump */
-  if (off != NULL) {
-    unsigned short u16 = (unsigned short)0xC000 |
-                         (unsigned short)(off->idx & 0x3FFF);
-    status             = ares__buf_append_be16(buf, u16);
-    if (status != ARES_SUCCESS) {
-      goto done;
-    }
-  }
-
-  /* Store pointer for future jumps as long as its not an exact match for
-   * a prior entry */
-  if (list != NULL && off != NULL && off->name_len != name_len && name_len > 0) {
-    status = ares__nameoffset_create(list, name /* not truncated copy! */, pos);
-    if (status != ARES_SUCCESS) {
-      goto done;
-    }
-  }
-
-  status = ARES_SUCCESS;
-
-done:
-  ares_free(labels);
-  return status;
-}
 
 static ares_status_t ares_dns_write_header(const ares_dns_record_t *dnsrec,
                                            ares__buf_t             *buf)
@@ -491,7 +141,7 @@ static ares_status_t ares_dns_write_questions(const ares_dns_record_t *dnsrec,
     }
 
     /* Name */
-    status = ares_dns_write_name(buf, namelist, ARES_TRUE, name);
+    status = ares__dns_name_write(buf, namelist, ARES_TRUE, name);
     if (status != ARES_SUCCESS) {
       return status;
     }
@@ -525,7 +175,7 @@ static ares_status_t ares_dns_write_rr_name(ares__buf_t         *buf,
     return ARES_EFORMERR;
   }
 
-  return ares_dns_write_name(buf, namelist, validate_hostname, name);
+  return ares__dns_name_write(buf, namelist, validate_hostname, name);
 }
 
 static ares_status_t ares_dns_write_rr_str(ares__buf_t         *buf,
@@ -1023,7 +673,7 @@ static ares_status_t ares_dns_write_rr(ares_dns_record_t *dnsrec,
 
     /* Name */
     status =
-      ares_dns_write_name(buf, namelist, ARES_TRUE, ares_dns_rr_get_name(rr));
+      ares__dns_name_write(buf, namelist, ARES_TRUE, ares_dns_rr_get_name(rr));
     if (status != ARES_SUCCESS) {
       return status;
     }
